refactor(crypto): move encrypted body validator to crypto services
Move EncryptedBodyValidator from infrastructure/validators to infrastructure/services/crypto/validators to better organize crypto-related code. Update imports and test mocks accordingly.
